rename file to match component name
