+++ src/DetailsView/reports/components/report-sections/result-section.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { NamedSFC } from '../../../../common/react/named-sfc';
 import { ResultSectionContent, ResultSectionContentProps } from './result-section-content';
 import { ResultSectionTitle, ResultSectionTitleProps } from './result-section-title';
-import { RulesWithInstancesDeps } from './rule-details-group';
+import { RulesWithInstancesDeps } from './rules-with-instances';
 
 export type ResultSectionDeps = RulesWithInstancesDeps;
