+++ src/DetailsView/reports/components/report-sections/rules-with-instances.tsx
